misc: add service-connect-config json example
Added a config json file that can be used to update the service
with regards to the Service Connect server related config. It requires
manual deregister for the older service task definition, so that the new
revision created as a result of the update to be used for running the
service.